StationとStationGroupの日付部分を分離、stationGroupInfoとStationInfoの処理を更新

+++ server/server.js
@@ -150,8 +150,6 @@ app.get("/api/stationGroup/:stationGroupCode", accessLog, (req, res, next) => {
     data = db.prepare(`
       SELECT
         StationGroups.*,
-        MAX(getDate) AS maxGetDate,
-        MAX(passDate) AS maxPassDate,
         Prefectures.name AS prefName
       FROM Stations
       INNER JOIN StationGroups
@@ -187,7 +185,6 @@ app.get("/api/stationsByGroupCode/:stationGroupCode", accessLog, (req, res, next
         Stations.*,
         StationGroups.stationName,
         StationGroups.kana,
-        StationGroups.date,
         Railways.railwayName,
         Railways.railwayColor,
         Companies.companyCode,
@@ -919,21 +916,8 @@ app.post("/api/login", accessLog, (req, res, next) => {
 
 // check
 app.get("/api/status", accessLog, (req, res, next) => {
-  const sessionId = req.cookies.sessionId;
-  if(!sessionId){
-    res.json({
-      auth: false,
-      userName: undefined,
-      userEmail: undefined,
-    })
-    return;
-  }
-  const userData = usersManager.status(sessionId);
-  res.json({
-    auth: userData !== undefined,
-    userName: userData?.userName,
-    userEmail: userData?.userEmail,
-  });
+  const userData = usersManager.getUserData(req);
+  res.json(userData);
 });
 
 // logout
@@ -960,6 +944,67 @@ app.get("/api/logout", accessLog, (req, res, next) => {
 ////////// History //////////
 
 
+// 駅の最新のアクセス日時を取得
+app.get("/api/latestStationHistory/:stationCode", accessLog, (req, res, next) => {
+  const code = +req.params.stationCode;
+  if(isNaN(code)){
+    next(new Error("Invalid input"));
+    return;
+  }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
+  let data;
+  try{
+    const stmt = db.prepare(`
+      SELECT date FROM LatestStationHistory
+      WHERE stationCode = ? AND state = ? AND userId = ?
+    `);
+    data = {
+      getDate: stmt.get(code, 0, userId)?.date ?? null,
+      passDate: stmt.get(code, 1, userId)?.date ?? null,
+    };
+  }catch(err){
+    console.error(err);
+    next(new Error("Server Error"));
+    return;
+  }
+  res.json(data);
+});
+
+// 駅グループの最新のアクセス日時を取得
+app.get("/api/latestStationGroupHistory/:stationGroupCode", accessLog, (req, res, next) => {
+  const code = +req.params.stationGroupCode;
+  if(isNaN(code)){
+    next(new Error("Invalid input"));
+    return;
+  }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
+  let data;
+  try{
+    const stmt = db.prepare(`
+      SELECT date FROM LatestStationGroupHistory
+      WHERE stationGroupCode = ? AND userId = ?
+    `);
+    data = {
+      date: stmt.get(code, userId)?.date ?? null,
+    };
+  }catch(err){
+    console.error(err);
+    next(new Error("Server Error"));
+    return;
+  }
+  res.json(data);
+});
+
 // 全体の乗降/通過の履歴を区間取得
 app.get("/api/stationHistory", accessLog, (req, res, next) => {
   const off = +req.query.off;
@@ -970,6 +1015,12 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let data;
   try{
     if(type === "station" && name !== ""){
@@ -990,6 +1041,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN StationGroups
           ON Stations.stationGroupCode = StationGroups.stationGroupCode
             AND StationGroups.stationName = ?
@@ -1002,7 +1054,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         ORDER BY StationHistory.date DESC
         LIMIT ?
         OFFSET ?
-      `).all(name, len, off);
+      `).all(userId, name, len, off);
     }else if(type === "railway" && name !== ""){
       data = db.prepare(`
         SELECT
@@ -1021,6 +1073,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN Railways
           ON Stations.railwayCode = Railways.railwayCode
             AND Railways.railwayName = ?
@@ -1033,7 +1086,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         ORDER BY StationHistory.date DESC
         LIMIT ?
         OFFSET ?
-      `).all(name, len, off);
+      `).all(userId, name, len, off);
     }else if(type === "company" && name !== ""){
       data = db.prepare(`
         SELECT
@@ -1052,6 +1105,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN StationGroups
           ON Stations.stationGroupCode = StationGroups.stationGroupCode
         INNER JOIN Railways
@@ -1064,7 +1118,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         ORDER BY StationHistory.date DESC
         LIMIT ?
         OFFSET ?
-      `).all(name, len, off);
+      `).all(userId, name, len, off);
     }else{
       data = db.prepare(`
         SELECT
@@ -1082,6 +1136,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN StationGroups
           ON Stations.stationGroupCode = StationGroups.stationGroupCode
         INNER JOIN Railways
@@ -1093,7 +1148,7 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
         ORDER BY date DESC
         LIMIT ?
         OFFSET ?
-      `).all(len, off);
+      `).all(userId, len, off);
     }
 
     data = data.map(station => insert_next_stations(station, station.stationCode));
@@ -1109,6 +1164,12 @@ app.get("/api/stationHistory", accessLog, (req, res, next) => {
 app.get("/api/stationHistoryCount", accessLog, (req, res, next) => {
   const name = req.query.name ?? "";
   const type = req.query.type;
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let data;
   try{
     if(type === "station" && name !== ""){
@@ -1116,34 +1177,38 @@ app.get("/api/stationHistoryCount", accessLog, (req, res, next) => {
         SELECT COUNT(*) AS count FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN StationGroups
           ON Stations.stationGroupCode = StationGroups.stationGroupCode
             AND StationGroups.stationName = ?
-      `).get(name);
+      `).get(userId, name);
     }else if(type === "railway" && name !== ""){
       data = db.prepare(`
         SELECT COUNT(*) AS count FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN Railways
           ON Stations.railwayCode = Railways.railwayCode
             AND Railways.railwayName = ?
-      `).get(name);
+      `).get(userId, name);
     }else if(type === "company" && name !== ""){
       data = db.prepare(`
         SELECT COUNT(*) AS count FROM StationHistory
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
+            AND StationHistory.userId = ?
         INNER JOIN Railways
           ON Stations.railwayCode = Railways.railwayCode
         INNER JOIN Companies
           ON Railways.companyCode = Companies.companyCode
             AND Companies.companyName = ?
-      `).get(name);
+      `).get(userId, name);
     }else{
       data = db.prepare(`
         SELECT COUNT(*) AS count FROM StationHistory
-      `).get();
+        WHERE StationHistory.userId = ?
+      `).get(userId);
     }
   }catch(err){
     console.error(err);
@@ -1155,6 +1220,12 @@ app.get("/api/stationHistoryCount", accessLog, (req, res, next) => {
 
 // 全体の乗降/通過の履歴と駅情報を取得
 app.get("/api/stationHistoryAndInfo", accessLog, (req, res, next) => {
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   // const off = req.query.off;
   // const len = req.query.len;
   // if(off === undefined || len === undefined){
@@ -1179,6 +1250,7 @@ app.get("/api/stationHistoryAndInfo", accessLog, (req, res, next) => {
       FROM StationHistory
       INNER JOIN Stations
         ON StationHistory.stationCode = Stations.stationCode
+          AND StationHistory.userId = ?
       INNER JOIN StationGroups
         ON Stations.stationGroupCode = StationGroups.stationGroupCode
       INNER JOIN Railways
@@ -1188,7 +1260,7 @@ app.get("/api/stationHistoryAndInfo", accessLog, (req, res, next) => {
       INNER JOIN Prefectures
         ON StationGroups.prefCode = Prefectures.code
       ORDER BY date DESC
-    `).all();
+    `).all(userId);
 
     data = data.map(station => insert_next_stations(station, station.stationCode));
   }catch(err){
@@ -1206,13 +1278,19 @@ app.get("/api/stationHistory/:stationCode", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let data;
   try{
     data = db.prepare(`
       SELECT * FROM StationHistory
-      WHERE stationCode = ?
+      WHERE stationCode = ? AND userId = ?
       ORDER BY date DESC
-    `).all(code);
+    `).all(code, userId);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1228,17 +1306,23 @@ app.get("/api/stationGroupHistory/:stationGroupCode", accessLog, (req, res, next
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
   let data;
   try{
     data = db.prepare(`
         SELECT
-          StationGroupHistory.*,
+        StationGroupHistory.stationGroupCode,
+          StationGroupHistory.date,
           2 AS state,
           '' AS railwayName,
           '' AS railwayColor,
           NULL AS stationCode
         FROM StationGroupHistory
-        WHERE stationGroupCode = ?
+        WHERE stationGroupCode = ? AND userId = ?
       UNION ALL
         SELECT
           Stations.stationGroupCode,
@@ -1251,10 +1335,11 @@ app.get("/api/stationGroupHistory/:stationGroupCode", accessLog, (req, res, next
         INNER JOIN Stations
           ON StationHistory.stationCode = Stations.stationCode
             AND Stations.stationGroupCode = ?
+            AND StationHistory.userId = ?
         INNER JOIN Railways
           ON Stations.railwayCode = Railways.railwayCode
       ORDER BY date DESC
-    `).all(code, code);
+    `).all(code, userId, code, userId);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1270,6 +1355,12 @@ app.get("/api/railwayProgress/:railwayCode", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let stationNum, getOrPassStationNum;
   try{
     stationNum = db.prepare(`
@@ -1282,7 +1373,8 @@ app.get("/api/railwayProgress/:railwayCode", accessLog, (req, res, next) => {
       INNER JOIN Stations
         ON StationHistory.stationCode = Stations.stationCode
           AND Stations.railwayCode = ?
-    `).get(code);
+          AND StationHistory.userId = ?
+    `).get(code, userId);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1298,6 +1390,12 @@ app.get("/api/companyProgress/:companyCode", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let stationNum, getOrPassStationNum;
   try{
     stationNum = db.prepare(`
@@ -1311,10 +1409,11 @@ app.get("/api/companyProgress/:companyCode", accessLog, (req, res, next) => {
       SELECT COUNT(DISTINCT(StationHistory.stationCode)) AS num FROM StationHistory
       INNER JOIN Stations
         ON StationHistory.stationCode = Stations.stationCode
+          AND StationHistory.userId = ?
       INNER JOIN Railways
         ON Stations.railwayCode = Railways.railwayCode
           AND Railways.companyCode = ?
-    `).get(code);
+    `).get(userId, code);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1330,6 +1429,12 @@ app.get("/api/prefProgress/:prefCode", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   let stationNum, getOrPassStationNum;
   try{
     stationNum = db.prepare(`
@@ -1343,10 +1448,11 @@ app.get("/api/prefProgress/:prefCode", accessLog, (req, res, next) => {
       SELECT COUNT(DISTINCT(StationHistory.stationCode)) AS num FROM StationHistory
       INNER JOIN Stations
         ON StationHistory.stationCode = Stations.stationCode
+          AND StationHistory.userId = ?
       INNER JOIN StationGroups
         ON Stations.stationGroupCode = StationGroups.stationGroupCode
           AND StationGroups.prefCode = ?
-    `).get(code);
+    `).get(userId, code);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1368,15 +1474,32 @@ app.get("/api/postStationDate", accessLog, (req, res, next) => {
     next(new RangeError("Invalid input"));
     return;
   }
-  const value_name = ["getDate", "passDate"][state];
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   try{
     db.prepare(
-      "INSERT INTO StationHistory VALUES(?, datetime(?), ?)"
-    ).run(code, date, state);
+      "INSERT INTO StationHistory VALUES(?, datetime(?), ?, ?)"
+    ).run(code, date, state, userId);
 
-    db.prepare(
-      `UPDATE Stations SET ${value_name} = MAX(IFNULL(${value_name}, 0), datetime(?)) WHERE stationCode = ?`
-    ).run(date, code);
+    const cnt = db.prepare(`
+      SELECT COUNT(*) AS cnt FROM LatestStationHistory
+      WHERE stationCode = ? AND state = ? AND userId = ?
+    `).get(code, state, userId).cnt;
+
+    if(cnt){
+      db.prepare(`
+        UPDATE LatestStationHistory SET date = MAX(IFNULL(date, 0), datetime(?))
+        WHERE stationCode = ? AND state = ? AND userId = ?
+      `).run(date, code, state, userId);
+    }else{
+      db.prepare(`
+        INSERT INTO LatestStationHistory VALUES(?, datetime(?), ?, ?)
+      `).run(code, date, state, userId);
+    }
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1393,13 +1516,32 @@ app.get("/api/postStationGroupDate", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
-  try{
-    db.prepare("INSERT INTO StationGroupHistory VALUES(?, datetime(?))")
-      .run(code, date);
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
 
-    db.prepare(
-      `UPDATE StationGroups SET date = MAX(IFNULL(date, 0), datetime(?)) WHERE stationGroupCode = ?`
-    ).run(date, code);
+  try{
+    db.prepare(`
+      INSERT INTO StationGroupHistory VALUES(?, datetime(?), ?)
+    `).run(code, date, userId);
+
+    const cnt = db.prepare(`
+      SELECT COUNT(*) AS cnt FROM LatestStationGroupHistory
+      WHERE stationGroupCode = ? AND userId = ?
+    `).get(code, userId).cnt;
+
+    if(cnt){
+      db.prepare(`
+        UPDATE LatestStationGroupHistory SET date = MAX(IFNULL(date, 0), datetime(?))
+        WHERE stationGroupCode = ? AND userId = ?
+      `).run(date, code, userId);
+    }else{
+      db.prepare(`
+        INSERT INTO LatestStationGroupHistory VALUES(?, datetime(?), ?)
+      `).run(code, date, userId);
+    }
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1421,20 +1563,26 @@ app.get("/api/deleteStationDate", accessLog, (req, res, next) => {
     next(new RangeError("Invalid input"));
     return;
   }
-  const value_name = ["getDate", "passDate"][state];
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   try{
     db.prepare(`
       DELETE FROM StationHistory
-      WHERE stationCode = ? AND date = datetime(?) AND state = ?
-    `).run(code, date, state);
+      WHERE stationCode = ? AND date = datetime(?) AND state = ? AND userId = ?
+    `).run(code, date, state, userId);
 
+    // 要素が何もなければNULLが入る
     db.prepare(`
-      UPDATE Stations SET ${value_name} = (
+      UPDATE LatestStationHistory SET date = (
         SELECT MAX(date) FROM StationHistory
-        WHERE stationCode = ? AND state = ?
+        WHERE stationCode = ? AND state = ? AND userId = ?
       )
-      WHERE stationCode = ? AND ${value_name} = datetime(?)
-    `).run(code, state, code, date);
+      WHERE stationCode = ? AND state = ? AND userId = ?
+    `).run(code, state, userId, code, state, userId);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
@@ -1451,19 +1599,26 @@ app.get("/api/deleteStationGroupState", accessLog, (req, res, next) => {
     next(new Error("Invalid input"));
     return;
   }
+  const userId = usersManager.getUserData(req).userId;
+  if(!userId){
+    next(new Error("Unauthorized"));
+    return;
+  }
+
   try{
     db.prepare(`
       DELETE FROM StationGroupHistory
-      WHERE stationGroupCode = ? AND date = datetime(?)
-    `).run(code, date);
+      WHERE stationGroupCode = ? AND date = datetime(?) AND userId = ?
+    `).run(code, date, userId);
 
+    // 要素が何もなければNULLが入る
     db.prepare(`
-      UPDATE StationGroups SET date = (
+      UPDATE LatestStationGroupHistory SET date = (
         SELECT MAX(date) FROM StationGroupHistory
-        WHERE stationGroupCode = ?
+        WHERE stationGroupCode = ? AND userId = ?
       )
-      WHERE stationGroupCode = ?
-    `).run(code, code);
+      WHERE stationGroupCode = ? AND userId = ?
+    `).run(code, userId, code, userId);
   }catch(err){
     console.error(err);
     next(new Error("Server Error"));
